refactor(sdk): address PR review comments
- Switch keygen/dummy.rs cuda branch to RootGpuProver (matches continuations)
- Move solidity constants, sol! macro, generate/verify/format functions
  from lib.rs into new solidity.rs module
- Delete static-verifier profile/.gitkeep
